requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
xgboost==2.0.3
imbalanced-learn==0.12.0
joblib==1.3.2
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.4
shap==0.45.0
